♻️ refactor: 공지 관련된 사용자 api 분리
