+++ src/main/java/com/real/backend/domain/user/controller/UserNoticeController.java
@@ -1,4 +1,4 @@
-package com.real.backend.domain.user;
+package com.real.backend.domain.user.controller;
 
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.real.backend.domain.user.dto.LoginResponseDTO;
 import com.real.backend.domain.user.dto.UserUnreadNoticeResponseDTO;
+import com.real.backend.domain.user.service.UserNoticeService;
 import com.real.backend.domain.user.service.UserService;
 import com.real.backend.response.DataResponse;
 import com.real.backend.response.StatusResponse;
@@ -21,16 +21,10 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
-public class UserController {
+public class UserNoticeController {
 
     private final UserService userService;
-
-    @GetMapping("/v1/users/info")
-    public DataResponse<?> getUserInfo(@CurrentSession Session session) {
-        LoginResponseDTO loginResponseDTO = userService.getUserInfo(session.getId());
-
-        return DataResponse.of(loginResponseDTO);
-    }
+    private final UserNoticeService userNoticeService;
 
     @PreAuthorize("!hasAnyAuthority('OUTSIDER')")
     @GetMapping("v1/users/notices/unread")
@@ -39,7 +33,7 @@ public class UserController {
         @RequestParam(value = "cursorStandard", required = false) String cursorStandard,
         @RequestParam(value = "limit", required = false, defaultValue = "10") int limit,
         @CurrentSession Session session) {
-        SliceDTO<UserUnreadNoticeResponseDTO> userUnreadNoticeResponseDTOList = userService.getNoticeListByCursor(
+        SliceDTO<UserUnreadNoticeResponseDTO> userUnreadNoticeResponseDTOList = userNoticeService.getNoticeListByCursor(
             cursorId, limit, cursorStandard, session.getId());
 
         return DataResponse.of(userUnreadNoticeResponseDTOList);
@@ -48,7 +42,7 @@ public class UserController {
     @PreAuthorize("!hasAnyAuthority('OUTSIDER')")
     @PostMapping("v1/users/notices/read")
     public StatusResponse readNotices(@CurrentSession Session session) {
-        userService.readNotices(session.getId());
+        userNoticeService.readNotices(session.getId());
         return StatusResponse.of(200, "안 읽은 공지들이 성공적으로 읽음 처리 되었습니다.");
     }
 }
